+++ build_runner/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.10.3
 
-- Use timings package for performance tracking.
+- Improve performance tracking and visualization using timing package
 
 ## 0.10.2